refactor(sign-up): rename folder that hold SignUpCubit from cubit to sign_up_cubit
