+++ lib/features/auth/presentation/manager/sign_up_cubit/sign_up_cubit.dart
