+++ lib/features/auth/presentation/views/widgets/already_have_an_account.dart
@@ -5,7 +5,7 @@ import 'package:nova_store_app/core/routing/routes_paths.dart';
 import 'package:nova_store_app/core/theme/app_colors.dart';
 import 'package:nova_store_app/core/theme/app_text_styles.dart';
 import 'package:nova_store_app/core/widgets/clickable_text.dart';
-import 'package:nova_store_app/features/auth/presentation/manager/cubit/sign_up_cubit.dart';
+import 'package:nova_store_app/features/auth/presentation/manager/sign_up_cubit/sign_up_cubit.dart';
 
 class AlreadyHaveAccount extends StatelessWidget {
   const AlreadyHaveAccount({
