+++ lib/features/auth/presentation/views/widgets/sign_up_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:nova_store_app/core/widgets/Custom_text_button.dart';
-import 'package:nova_store_app/features/auth/presentation/manager/cubit/sign_up_cubit.dart';
+import 'package:nova_store_app/features/auth/presentation/manager/sign_up_cubit/sign_up_cubit.dart';
 
 class SignUpButton extends StatelessWidget {
   const SignUpButton({
